+++ CHANGELOG.md
@@ -19,6 +19,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Runtime on ESM with Jest tests remaining with the babel transpiler. Yeoman constrains us to commonJS for the library itself ([#29](https://github.com/LaunchAcademy/generator-engage/issues/29))
 - Create a `Dependency` abstraction to encapsulate both npm package and correlating version
 
+### Removed
+
+- Sequelize configuration is no longer an option ([#48](https://github.com/LaunchAcademy/generator-engage/issues/48))
+
 ## [0.1.2](https://github.com/LaunchAcademy/generator-engage/releases/tag/v0.1.2)
 
 ### Added